Only add FILE.TXT to ESP once, at the end.
Disk image size was not being written or read correctly
from FILE.TXT in the ESP from UEFI, overwriting or adding
twice was not working to correctly add the image size.

Only need to write the file once to the ESP anyway, can do that
at the end of main() which fixes this issue.