Fix ESP size > 32MB and adding files to ESP > 32MiB
Fix the BPB_FATSz32 value in the ESP's FAT32 VBR, to be
more in line with the fatgen103.doc from M$. It should
contain the rough amount of disk LBAs or sectors needed
to contain clusters (4 byte values, 512 byte LBA = 128 clusters)
for the size of the filesystem.

Fix cluster mirroring when adding a file to the FAT tables.
The next free cluster was not being reset between fseek() calls
to each FAT, which would be corrupting the FAT table mirroring,
that could cause random issues depending on which FAT table is
used by firmware or other filesystem drivers.